Implemented "stop suggesting" feature
When a user press one of the pre-defined keys (for example, "Enter" or "Space"),
asuggest stops auto-completing and then the space or another pre-defined symbol
is appended to the text.

New asuggest options:

- endingSymbols: string which will be appended at the end of the suggested text
  after this kind of breaking. Space is the default one.
- stopSuggestionKeys: array of keycodes (int), each of which defines a key
  which stops suggesting. Codes of space and return are used by default.

In order to prevent such behaviour it's enough to pass an empty array
for a "stopSuggestionKeys" argument.

There is a set of constants in $.asuggestKeys variable, which simplifies the
definition of stopSuggestionKeys array.

Closes #2